Add choice of Kahn's or DFS topo sort in TopologicalSort
